package updates and layer tests

+++ layer_builds.sh
@@ -10,10 +10,10 @@ py_version=$(python --version | grep -o -E '[0-9]+[.][0-9]+')
 py_ver=$(echo "py${py_version}" | sed -e 's/\.//g')
 
 # The Makefile provides values for these environment variables:
-#LIB_NAME
-#LIB_VERSION
-#LIB_REPO
-#LIB_PACKAGE
+test -n "${LIB_NAME}" || crash "LIB_NAME env-var must be defined"
+test -n "${LIB_VERSION}" || crash "LIB_VERSION env-var must be defined"
+test -n "${LIB_PACKAGE}" || crash "LIB_PACKAGE env-var must be defined"
+
 LAYER_PREFIX="${py_ver}-${LIB_NAME}-${LIB_VERSION}"
 
 ZIP_PATH=${ZIP_PATH:-'/tmp'}
@@ -21,15 +21,14 @@ ZIP_PATH=$(readlink -f "$ZIP_PATH")
 ZIP_TMP="${ZIP_PATH}/${py_ver}_lambda_layer.zip"
 
 REQUIREMENTS_FILE="${ZIP_PATH}/requirements.txt"
+test -f "$REQUIREMENTS_FILE" || crash "ERROR: there is no $REQUIREMENTS_FILE"
 
-if [ ! -f "$REQUIREMENTS_FILE" ]; then
-  echo "ERROR: there is no $REQUIREMENTS_FILE"
-  exit 1
-fi
-
-crash() {
-  echo "OOPS - something went wrong!"
-  exit 1
+move_zip () {
+  zip_lib=$1
+  mv "${ZIP_TMP}" "${zip_lib}" || crash "Failed to create ${zip_lib}"
+  echo "Created ${zip_lib}"
+  echo
+  echo
 }
 
 #
@@ -38,24 +37,20 @@ crash() {
 
 cp -p "${ZIP_PATH}/requirements.txt" /tmp/requirements.txt
 create_layer_zip
-mv "${ZIP_TMP}" "${ZIP_PATH}/${LIB_PACKAGE}" || crash
-echo "Created ${ZIP_PATH}/${LIB_PACKAGE}"
-echo
-echo
+move_zip "${ZIP_PATH}/${LIB_PACKAGE}"
 
 #
 # Package Optional Extras
 #
 
 for f in "${ZIP_PATH}"/requirements_*.txt; do
-  # shellcheck disable=SC2001
-  extra=$(echo "${f}" | sed 's/.*requirements_\(.*\)\.txt/\1/')
-  cp -p "${f}" /tmp/requirements.txt
-  create_layer_zip
-  mv "${ZIP_TMP}" "${ZIP_PATH}/${LAYER_PREFIX}-${extra}.zip" || crash
-  echo "Created ${ZIP_PATH}/${LAYER_PREFIX}-${extra}.zip"
-  echo
-  echo
+  if [ -f "${f}" ]; then
+    # shellcheck disable=SC2001
+    extra=$(echo "${f}" | sed 's/.*requirements_\(.*\)\.txt/\1/')
+    cp -p "${f}" /tmp/requirements.txt
+    create_layer_zip
+    move_zip "${ZIP_PATH}/${LAYER_PREFIX}-${extra}.zip"
+  fi
 done
 
 
@@ -78,21 +73,23 @@ cp -rf "${ZIP_PATH}"/project-no-deps/* "$package_dst/"
 echo "Zipping packages for lambda layer..."
 layer_zip="${ZIP_PATH}/${LAYER_PREFIX}-nodeps.zip"
 rm -f "${layer_zip}"
-pushd "${package_dir}" >/dev/null || crash
+pushd "${package_dir}" >/dev/null || crash "Failed to pushd ${package_dir}"
 
 zip -qr9 --compression-method deflate --symlinks "${layer_zip}" python
 zip -qr9 --compression-method deflate --symlinks "${ZIP_PATH}/${LIB_PACKAGE}" python
 for f in "${ZIP_PATH}"/requirements_*.txt; do
-  # shellcheck disable=SC2001
-  extra=$(echo "${f}" | sed 's/.*requirements_\(.*\)\.txt/\1/')
-  extra_zip="${ZIP_PATH}/${LAYER_PREFIX}-${extra}.zip"
-  zip -rq9 --compression-method deflate --symlinks "${extra_zip}" python
+  if [ -f "${f}" ]; then
+    # shellcheck disable=SC2001
+    extra=$(echo "${f}" | sed 's/.*requirements_\(.*\)\.txt/\1/')
+    extra_zip="${ZIP_PATH}/${LAYER_PREFIX}-${extra}.zip"
+    zip -qr9 --compression-method deflate --symlinks "${extra_zip}" python
+  fi
 done
 
-ls "${layer_zip}" >/dev/null || crash
-unzip -q -t "${layer_zip}" || crash
+ls "${layer_zip}" >/dev/null || crash "Failed to find ${layer_zip}"
+unzip -q -t "${layer_zip}" || crash "Failed to test ${layer_zip}"
 echo "created ${layer_zip}"
-popd >/dev/null || exit 1
+popd >/dev/null || crash "Failed to popd from ${package_dir}"
 rm -rf "$package_dir"
 echo
 echo
